tls: allow tls.rootCertificates to be set
Modifies the tls.rootCertificates property so that it is settable. PEM-formatted certificates set to the property are loaded into a new Node.js root certificate store that is used for all subsequent TLS requests and peer certificate validation.

Allows full programmatic control of a Node.js process' root certificate store.

Fixes: https://github.com/nodejs/node/issues/20432
Refs: https://github.com/nodejs/node/issues/27079